instagram_content_ai/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
asyncer==0.0.7
attrs==25.1.0
backoff==2.2.1
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chainlit==2.5.5
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.2
CurrencyConverter==0.18.3
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
flexcache==0.3
flexparser==0.4
fonttools==4.56.0
forex-python==1.8
frozenlist==1.6.2
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
Jinja2==3.1.5
jiter==0.10.0
joblib==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
Lazify==0.4.0
literalai==0.1.201
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
mcp==1.9.4
mdurl==0.1.2
monotonic==1.6
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.28.0
nest-asyncio==1.6.0
numpy==2.2.3
openai==1.84.0
openpyxl==3.1.5
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-alephalpha==0.40.9
opentelemetry-instrumentation-anthropic==0.40.9
opentelemetry-instrumentation-bedrock==0.40.9
opentelemetry-instrumentation-chromadb==0.40.9
opentelemetry-instrumentation-cohere==0.40.9
opentelemetry-instrumentation-crewai==0.40.9
opentelemetry-instrumentation-google-generativeai==0.40.9
opentelemetry-instrumentation-groq==0.40.9
opentelemetry-instrumentation-haystack==0.40.9
opentelemetry-instrumentation-lancedb==0.40.9
opentelemetry-instrumentation-langchain==0.40.9
opentelemetry-instrumentation-llamaindex==0.40.9
opentelemetry-instrumentation-logging==0.52b1
opentelemetry-instrumentation-marqo==0.40.9
opentelemetry-instrumentation-mcp==0.40.9
opentelemetry-instrumentation-milvus==0.40.9
opentelemetry-instrumentation-mistralai==0.40.9
opentelemetry-instrumentation-ollama==0.40.9
opentelemetry-instrumentation-openai==0.40.9
opentelemetry-instrumentation-pinecone==0.40.9
opentelemetry-instrumentation-qdrant==0.40.9
opentelemetry-instrumentation-replicate==0.40.9
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-sagemaker==0.40.9
opentelemetry-instrumentation-sqlalchemy==0.52b1
opentelemetry-instrumentation-threading==0.52b1
opentelemetry-instrumentation-together==0.40.9
opentelemetry-instrumentation-transformers==0.40.9
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-vertexai==0.40.9
opentelemetry-instrumentation-watsonx==0.40.9
opentelemetry-instrumentation-weaviate==0.40.9
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-semantic-conventions-ai==0.4.9
opentelemetry-util-http==0.52b1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
Pint==0.24.4
platformdirs==4.3.6
plotly==6.0.0
posthog==3.25.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.2
python-multipart==0.0.18
python-socketio==5.13.0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
simple-websocket==1.1.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==2.3.6
starlette==0.41.3
streamlit==1.42.2
syncer==2.0.3
tenacity==9.0.0
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traceloop-sdk==0.40.9
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
uptrace==1.31.0
uritemplate==4.1.1
urllib3==2.3.0
uv==0.7.11
uvicorn==0.34.3
watchdog==6.0.0
watchfiles==0.20.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zipp==3.23.0
